Start work on implementing [2006].
I'm currently stuck implementing the interleaving permutation.
It seems that I'm running into numeric errors since 1/gamma is
not accurately rendered, so for example if gamma = 0.3, then
1/gamma is 3.33 recurring, but when I multiply that by 3
and take an int/floor, I get 9 instead of the expected value 10.

This seems to be only a problem in calculating the rho value for
the redundant symbols.