Change structures to use void* where there's a choice of field to operate on.
Turn the solver routine into a macro.

Use aliases to data structures in d so that all pointer arithmetic
is done with the correct gfx_t type. (so use alpha instead of code_size
and blocksyms instead of blocksize).

Code continues to work with gf8 fields, but I need to write a new
encoder before I can test whether it's working for other fields.