Working with WS echo example in Guides::Cookbook as a base.
I want to have nodes in a cluster respond to two types of message:

* split and send some number of shares to other node(s)
* receive share from some node and store it

I want to use regular sockets for the share transport, so the
first step is to write the receiving side. This script will
take a file name/transaction ID/whatever (it doesn't matter which)
and set up a new listening port, then report back to the ws client
on the port number.

Later on, to simulate receiving a share, when the client receives
a file over the port, it calculates a SHA sum of the data and
reports it back on the ws channel (with the same filename/trans-
action id/etc.).